Move types to types.ts, reorder functions
